examples/health_assistant/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.1
langchain==0.0.267
langchain-openai==0.0.2.post1
openai
websocket-server==0.6.4
websockets==11.0.3 